+++ src/frontend/menu/settings/gamepad/preview/DpadCustom.qml
@@ -45,8 +45,6 @@ Item {
         anchors.horizontalCenterOffset: (parent.width/2) * -(pressAngle/1000)
 
         source: initialImage
-        brightness: 0.5
-        contrast: 0.5
         transform: Rotation { origin.x: initialImage.width/2; origin.y: initialImage.height/2; axis { x: 0; y: 1; z: 0 } angle: -pressAngle }
     }
 
@@ -73,8 +71,6 @@ Item {
         anchors.horizontalCenterOffset: (parent.width/2) * (pressAngle/1000)
 
         source: initialImage
-        brightness: 0.5
-        contrast: 0.5
         transform: Rotation { origin.x: initialImage.width/2; origin.y: initialImage.height/2; axis { x: 0; y: 1; z: 0 } angle: pressAngle }
     }
 
@@ -101,8 +97,6 @@ Item {
         anchors.verticalCenterOffset: (parent.height/2) * -(pressAngle/1000)
 
         source: initialImage
-        brightness: 0.5
-        contrast: 0.5
         transform: Rotation { origin.x: initialImage.width/2; origin.y: initialImage.height/2; axis { x: 1; y: 0; z: 0 } angle: pressAngle }
     }
 
@@ -128,8 +122,6 @@ Item {
         anchors.verticalCenterOffset: (parent.height/2) * (pressAngle/1000)
 
         source: initialImage
-        brightness: 0.5
-        contrast: 0.5
         transform: Rotation { origin.x: initialImage.width/2; origin.y: initialImage.height/2; axis { x: 1; y: 0; z: 0 } angle: -pressAngle }
     }